Scriptor-Server/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
Flask-RESTful==0.3.10
groq==0.11.0
python-dotenv==1.0.1